Enabled no panic minus literal.
commit-id:9e8aa88a
